fix msg

+++ database/migrations/Social/2024_12_21_140530_message_total_purchse.php
@@ -8,7 +8,7 @@ return new class () extends Migration {
     public function up(): void
     {
         Schema::table('messages', function (Blueprint $table) {
-            $table->tinyInteger('total_purchased')->after('total_shared')->default(0)->index('total_purchased');
+            $table->integer('total_purchased')->after('total_shared')->default(0)->index('total_purchased');
         });
     }
